done with setting logic

+++ app/src/main/java/com/android/greenapp/data/network/dto/coins/CoinRecord.kt
@@ -3,7 +3,7 @@ package com.android.greenapp.data.network.dto.coins
 data class CoinRecord(
     val coin: Coin,
     val coinbase: Boolean,
-    val confirmed_block_index: Int,
+    val confirmed_block_index: Long,
     val spent: Boolean,
     val spent_block_index: Int,
     val timestamp: Long
